start splitting into new and old
